# perf(sql): improve performance of length(VARCHAR)