Uses a SWAR technique to count the UTF-8 continuation bytes and subtract that from the total size of the VARCHAR.

Benchmark results:

```plain
STRING:

select * from (
    select title t1, title2 t2 from hits
) where length(t1) > 10;


VARCHAR:

select * from (
    select title t1, title2 t2 from hits
) where length(t2) > 10;


master:

STRING
10.8 s

VARCHAR
12 s


branch:

STRING
10.8 s

VARCHAR
9.5 s
```