@puzpuzpuz 2024-04-25
nit: I'd also add some rnd function-based tests to `LengthFunctionFactoryTest`. Currently, it only tests for a few short varchar values, so having something like `select count() from x where length(varchar_col) != length(str_col)` equals 0 may be a good idea.