@ideoma 2024-04-25
#### [PR Coverage check]

:heart_eyes: **pass** : 20 / 20 (100.00%)




#### file detail

|   |path|covered line|new line|coverage|
|----|----|----|----|----|
|:large_blue_circle:|io/questdb/std/str/Utf8s.java|19|19|100.00%|
|:large_blue_circle:|io/questdb/griffin/engine/functions/str/LengthVarcharFunctionFactory.java|1|1|100.00%|